added support for ogg files, code cleanup

+++ Apollo Soundboard/Forms/Downloading.cs
@@ -14,7 +14,7 @@ namespace Apollo_Soundboard.Forms
         private async void Downloading_VisibleChanged(object sender, EventArgs e)
         {
             await FFmpegDownloader.GetLatestVersion(FFmpegVersion.Official,
-                new Progress<ProgressInfo>(percent => progressBar1.Value = (int)(100 * ((float)percent.DownloadedBytes / (float)percent.TotalBytes))
+                new Progress<ProgressInfo>(percent => progressBar1.Value = (int)(100 * (percent.DownloadedBytes / (float)percent.TotalBytes))
                 ));
             Close();
         }
